refactor: Combine patches together
