+++ patches/src/main/kotlin/app/morphe/patches/youtube/interaction/seekbar/LivestreamDvrPatch.kt
@@ -7,6 +7,7 @@
 
 package app.morphe.patches.youtube.interaction.seekbar
 
+import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
 import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
@@ -31,7 +32,8 @@ val livestreamDvrPatch = bytecodePatch(
 
     execute {
         PreferenceScreen.SEEKBAR.addPreferences(
-            SwitchPreference("morphe_livestream_dvr")
+            SwitchPreference("morphe_livestream_dvr"),
+            SwitchPreference("morphe_expand_livestream_dvr_duration")
         )
 
         VideoStreamingDataAllowSeekingFingerprint.method.apply {
@@ -46,6 +48,19 @@ val livestreamDvrPatch = bytecodePatch(
                     """
                 )
             }
+        }
+
+        FormatStreamModelMaxDvrDurationFingerprint.method.apply {
+            val index = FormatStreamModelMaxDvrDurationFingerprint.instructionMatches.last().index
+            val register = getInstruction<OneRegisterInstruction>(index).registerA
+
+            addInstructions(
+                index,
+                """
+                    invoke-static { v$register, v${register + 1} }, $EXTENSION_CLASS_DESCRIPTOR->overrideMaxDvrDurationSec(D)D
+                    move-result-wide v$register
+                """
+            )
         }
     }
 }
