+++ patches/src/main/kotlin/app/morphe/patches/youtube/interaction/seekbar/SeekbarPatch.kt
@@ -19,7 +19,6 @@ val seekbarPatch = bytecodePatch(
         enableTapToSeekPatch,
         hideSeekbarPatch,
         livestreamDvrPatch,
-        expandLivestreamDvrDurationPatch,
     )
 
     compatibleWith(COMPATIBILITY_YOUTUBE)
